requirements.txt
asgiref==3.2.3
Django==3.0
django-crispy-forms==1.8.1
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
gunicorn